update: Move 'Read Next' to Pub footer
