+++ client/containers/Pub/usePageContext.js
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import { PageContext } from 'components/PageWrapper/PageWrapper';
+
+export const usePageContext = () => {
+	return useContext(PageContext);
+};
