+++ stories/containers/Pub/PubBottom/pubBottomStories.js
@@ -50,6 +50,7 @@ const pubData = {
 	activeBranch: {
 		id: 'ff116d8b-c3e5-4472-ab1a-8041ec1cc842',
 	},
+	collectionPubs: [],
 };
 
 storiesOf('containers/Pub/PubDocument/PubBottom/PubBottom', module)
